feat: optimise electron code
