+++ .gitignore
@@ -3,3 +3,6 @@ weights/model.onnx
 venv/
 __pycache__/
 node_modules/
+electron/node_modules/
+electron/python-embedded/
+electron/dist/
