+++ electron/app_is_packaged.js
@@ -1,0 +1,3 @@
+// app_is_packaged.js
+const { app } = require('electron');
+console.log('App is packaged:', app.isPackaged);
